+++ graphene_mongo/asynchronous/fields.py
@@ -4,18 +4,15 @@ from functools import partial
 from itertools import filterfalse
 from typing import Coroutine
 
-import bson
-from bson import DBRef, ObjectId
 import graphene
+from bson import DBRef
 from graphene import Context
 from graphene.relay import ConnectionField
-from graphene.utils.str_converters import to_snake_case
 from graphql import GraphQLResolveInfo
 from graphql_relay import cursor_to_offset, from_global_id
 import mongoengine
 from mongoengine import AsyncQuerySet, QuerySet
 from promise import Promise
-import pymongo
 from pymongo.errors import OperationFailure
 
 from ..synchronous.fields import MongoengineConnectionField
@@ -24,21 +21,11 @@ from ..base.utils import (
     ExecutorEnum,
     connection_from_iterables,
     find_skip_and_limit,
-    get_document,
-    get_model_reference_fields,
-    get_query_fields,
-    get_related_field_filter_args,
-    get_select_related_paths,
     has_page_info,
 )
 
-PYMONGO_VERSION = tuple(pymongo.version_tuple[:2])
-
 
 class AsyncMongoengineConnectionField(MongoengineConnectionField):
-    def __init__(self, type, *args, **kwargs):
-        super(AsyncMongoengineConnectionField, self).__init__(type, *args, **kwargs)
-
     @property
     def executor(self):
         return ExecutorEnum.ASYNC
@@ -57,52 +44,19 @@ class AsyncMongoengineConnectionField(MongoengineConnectionField):
         return _type._meta.connection
 
     @property
-    def fields(self):
-        return super(AsyncMongoengineConnectionField, self).fields
+    def registry(self):
+        return getattr(self.node_type._meta, "registry", get_global_async_registry())
+
+    def _qs_accessor(self, model):
+        return model.aobjects
 
     def get_queryset(
         self, model, info, required_fields=None, skip=None, limit=None, **args
     ) -> AsyncQuerySet:
         if required_fields is None:
             required_fields = list()
-
         if args:
-            reference_fields = get_model_reference_fields(self.model)
-            hydrated_references = {}
-            for arg_name, arg in args.copy().items():
-                if arg_name in reference_fields and not isinstance(
-                    arg, mongoengine.base.metaclasses.TopLevelDocumentMetaclass
-                ):
-                    try:
-                        reference_obj = reference_fields[arg_name].document_type(
-                            pk=from_global_id(arg)[1]
-                        )
-                    except TypeError:
-                        reference_obj = reference_fields[arg_name].document_type(pk=arg)
-                    hydrated_references[arg_name] = reference_obj
-                elif arg_name in self.model._fields_ordered and isinstance(
-                    getattr(self.model, arg_name),
-                    mongoengine.fields.GenericReferenceField,
-                ):
-                    try:
-                        reference_obj = get_document(
-                            self.registry.get_type_for_model_string(from_global_id(arg)[0])
-                        )(pk=from_global_id(arg)[1])
-                    except TypeError:
-                        reference_obj = get_document(arg["_cls"])(pk=arg["_ref"].id)
-                    hydrated_references[arg_name] = reference_obj
-                elif "__near" in arg_name and isinstance(
-                    getattr(self.model, arg_name.split("__")[0]),
-                    mongoengine.fields.PointField,
-                ):
-                    location = args.pop(arg_name, None)
-                    hydrated_references[arg_name] = location["coordinates"]
-                    if (arg_name.split("__")[0] + "__max_distance") not in args:
-                        hydrated_references[arg_name.split("__")[0] + "__max_distance"] = 10000
-                elif arg_name == "id":
-                    hydrated_references["id"] = from_global_id(args.pop("id", None))[1]
-            args.update(hydrated_references)
-
+            self._hydrate_args(args)
         if self._get_queryset:
             queryset_or_filters = self._get_queryset(model, info, **args)
             if isinstance(queryset_or_filters, mongoengine.AsyncQuerySet):
@@ -111,27 +65,17 @@ class AsyncMongoengineConnectionField(MongoengineConnectionField):
                 raise TypeError(
                     "AsyncMongoengineConnectionField only accepts AsyncQuerySet in get_queryset(...)"
                 )
-            else:
-                args.update(queryset_or_filters)
-        queried_fields = get_query_fields(info) if isinstance(info, GraphQLResolveInfo) else {}
-        related = get_select_related_paths(model, queried_fields)
-        related_filter = get_related_field_filter_args(info, model) if isinstance(info, GraphQLResolveInfo) else {}
-
-        qs = model.aobjects(**args).only(*required_fields).order_by(self.order_by)
-        if related:
-            qs = qs.select_related(*related)
-            for field_name, field_filter in related_filter.items():
-                if field_name in related:
-                    qs = qs.filter(**{f"{field_name}__{k}": v for k, v in field_filter.items()})
+            args.update(queryset_or_filters)
+        qs = self._apply_select_related(
+            self._qs_accessor(model)(**args).only(*required_fields).order_by(self.order_by),
+            model,
+            info,
+        )
         if limit is not None:
             return qs.skip(skip if skip else 0).limit(limit)
         elif skip is not None:
             return qs.skip(skip)
         return qs
-
-    @property
-    def registry(self):
-        return getattr(self.node_type._meta, "registry", get_global_async_registry())
 
     async def default_resolver(self, _root, info, required_fields=None, resolved=None, **args):
         if required_fields is None:
@@ -140,35 +84,13 @@ class AsyncMongoengineConnectionField(MongoengineConnectionField):
         for key, value in dict(args).items():
             if value is None:
                 del args[key]
-        if _root is not None and not resolved:
-            field_name = to_snake_case(info.field_name)
-            if not hasattr(_root, "_fields_ordered"):
-                if isinstance(getattr(_root, field_name, []), list):
-                    args["pk__in"] = [r.id for r in getattr(_root, field_name, [])]
-            elif field_name in _root._fields_ordered and not (
-                isinstance(_root._fields[field_name].field, mongoengine.EmbeddedDocumentField)
-                or isinstance(
-                    _root._fields[field_name].field,
-                    mongoengine.GenericEmbeddedDocumentField,
-                )
-            ):
-                raw = getattr(_root, field_name, [])
-                if raw is not None:
-                    first = next(iter(raw), None)
-                    if isinstance(first, mongoengine.Document):
-                        # Pre-loaded by select_related; the filter was already pushed into
-                        # the $lookup sub-pipeline by get_queryset via filter(**related_filter).
-                        # Return the pre-loaded (already filtered) list directly.
-                        resolved = list(raw)
-                        for k in [k for k in args if k != "id"]:
-                            args.pop(k)
-                    else:
-                        args["pk__in"] = [r.id for r in raw]
+
+        field_name, resolved = self._prepare_resolver_inputs(_root, info, args, resolved)
 
         _id = args.pop("id", None)
-
         if _id is not None:
             args["pk"] = from_global_id(_id)[-1]
+
         iterables = []
         list_length = 0
         skip = 0
@@ -209,7 +131,9 @@ class AsyncMongoengineConnectionField(MongoengineConnectionField):
                     items = await _base_query.limit(limit)
                     has_next_page = (
                         (
-                            len(await _base_query.skip(skip + limit).only("id").limit(1).to_list())
+                            len(
+                                await _base_query.skip(skip + limit).only("id").limit(1).to_list()
+                            )
                             != 0
                         )
                         if requires_page_info
@@ -240,10 +164,10 @@ class AsyncMongoengineConnectionField(MongoengineConnectionField):
                         args["pk__in"] = args["pk__in"][skip : skip + limit]
                     elif skip:
                         args["pk__in"] = args["pk__in"][skip:]
-                    iterables = self.get_queryset(self.model, info, required_fields, **args)
-                    iterables = await iterables.to_list()
+                    iterables = await self.get_queryset(
+                        self.model, info, required_fields, **args
+                    ).to_list()
                 else:
-                    # If there is no ids to fetch, No need of DB call
                     iterables = []
 
                 list_length = len(iterables)
@@ -258,29 +182,12 @@ class AsyncMongoengineConnectionField(MongoengineConnectionField):
                 or args
                 or isinstance(getattr(_root, field_name, []), AsyncMongoengineConnectionField)
             ):
-                args_copy = args.copy()
-                for key in args.copy():
-                    if key not in self.model._fields_ordered:
-                        args_copy.pop(key)
-                    elif isinstance(
-                        getattr(self.model, key), mongoengine.fields.ReferenceField
-                    ) or isinstance(
-                        getattr(self.model, key),
-                        mongoengine.fields.GenericReferenceField,
-                    ):
-                        if not isinstance(args_copy[key], ObjectId):
-                            _from_global_id = from_global_id(args_copy[key])[1]
-                            if bson.objectid.ObjectId.is_valid(_from_global_id):
-                                args_copy[key] = ObjectId(_from_global_id)
-                            else:
-                                args_copy[key] = _from_global_id
-                    elif isinstance(getattr(self.model, key), mongoengine.fields.EnumField):
-                        if getattr(args_copy[key], "value", None):
-                            args_copy[key] = args_copy[key].value
+                args_copy = self._build_args_copy(args)
 
                 if first is None and last is None and before is None and after is None:
-                    iterables = self.get_queryset(self.model, info, required_fields, **args)
-                    iterables = await iterables.to_list()
+                    iterables = await self.get_queryset(
+                        self.model, info, required_fields, **args
+                    ).to_list()
                     list_length = len(iterables)
                     if isinstance(info, GraphQLResolveInfo):
                         if not info.context:
@@ -294,13 +201,15 @@ class AsyncMongoengineConnectionField(MongoengineConnectionField):
                         count = await self.model.aobjects(**args_copy).count()
                     if not needs_count or count != 0:
                         skip, limit = find_skip_and_limit(
-                            first=first, after=after, last=last, before=before,
+                            first=first,
+                            after=after,
+                            last=last,
+                            before=before,
                             count=count if needs_count else None,
                         )
-                        iterables = self.get_queryset(
+                        iterables = await self.get_queryset(
                             self.model, info, required_fields, skip, limit, **args
-                        )
-                        iterables = await iterables.to_list()
+                        ).to_list()
                         list_length = len(iterables)
                         if isinstance(info, GraphQLResolveInfo):
                             if not info.context:
@@ -310,7 +219,6 @@ class AsyncMongoengineConnectionField(MongoengineConnectionField):
                             )
 
         elif _root is not None:
-            field_name = to_snake_case(info.field_name)
             items = getattr(_root, field_name, [])
             count = len(items)
             skip, limit = find_skip_and_limit(
@@ -322,8 +230,7 @@ class AsyncMongoengineConnectionField(MongoengineConnectionField):
                 has_next_page = (skip + limit) < len(_base_query) if requires_page_info else False
             elif skip:
                 items = items[skip:]
-            iterables = items
-            iterables = await iterables.to_list()
+            iterables = await items.to_list()
             list_length = len(iterables)
 
         if requires_page_info and count:
@@ -352,16 +259,7 @@ class AsyncMongoengineConnectionField(MongoengineConnectionField):
             if value is None:
                 del args[key]
 
-        required_fields = list()
-
-        for field in self.required_fields:
-            if field in self.model._fields_ordered:
-                required_fields.append(field)
-
-        for field in get_query_fields(info):
-            if to_snake_case(field) in self.model._fields_ordered:
-                required_fields.append(to_snake_case(field))
-
+        required_fields = self._collect_required_fields(info)
         args_copy = args.copy()
 
         if not bool(args) or not is_partial:
@@ -395,51 +293,22 @@ class AsyncMongoengineConnectionField(MongoengineConnectionField):
                         self.model, info, required_fields, **args_copy
                     )
 
-            # XXX: Filter nested args
             resolved = resolver(root, info, **args)
             if isinstance(resolved, Coroutine):
                 resolved = await resolved
             if resolved is not None:
-                # if isinstance(resolved, Coroutine):
-                #     resolved = await resolved
                 if isinstance(resolved, list):
                     if resolved == list():
                         return resolved
                     elif not isinstance(resolved[0], DBRef):
                         return resolved
                     else:
-                        return await self.default_resolver(root, info, required_fields, **args_copy)
+                        return await self.default_resolver(
+                            root, info, required_fields, **args_copy
+                        )
                 elif isinstance(resolved, QuerySet):
                     args.update(resolved._query)
-                    args_copy = args.copy()
-                    for arg_name, arg in args.copy().items():
-                        if "." in arg_name or arg_name not in self.model._fields_ordered + (
-                            "first",
-                            "last",
-                            "before",
-                            "after",
-                        ) + tuple(self.filter_args.keys()):
-                            args_copy.pop(arg_name)
-                            if arg_name == "_id" and isinstance(arg, dict):
-                                operation = list(arg.keys())[0]
-                                args_copy["pk" + operation.replace("$", "__")] = arg[operation]
-                            if not isinstance(arg, ObjectId) and "." in arg_name:
-                                if isinstance(arg, dict):
-                                    operation = list(arg.keys())[0]
-                                    args_copy[
-                                        arg_name.replace(".", "__") + operation.replace("$", "__")
-                                    ] = arg[operation]
-                                else:
-                                    args_copy[arg_name.replace(".", "__")] = arg
-                            elif "." in arg_name and isinstance(arg, ObjectId):
-                                args_copy[arg_name.replace(".", "__")] = arg
-                        else:
-                            operations = ["$lte", "$gte", "$ne", "$in"]
-                            if isinstance(arg, dict) and any(op in arg for op in operations):
-                                operation = list(arg.keys())[0]
-                                args_copy[arg_name + operation.replace("$", "__")] = arg[operation]
-                                del args_copy[arg_name]
-
+                    args_copy = self._transform_qs_args(args, args.copy())
                     return await self.default_resolver(
                         root, info, required_fields, resolved=resolved, **args_copy
                     )
